Close sockets on receiving exceptions during SSL runHandshake phase.
Without this, it may leak sockets that hold the connection until the server decides to kill it based on its idle channel configuration.
Possible reasons this may arise are data mangling over the wire, bad certificate fingerprint,etc - these issues result in SSL handshake failure at the 'unwrap' phase of OpenSSLEngine, which gets thrown all the way up without the OpenSSLSocket getting a chance to close it gracefully.

If enough sockets are leaked, it may result in serious issues like denial of service.